Refactor VertexEmitter into a QuadEmitter
Somehow this hits a happier path in the JVM. I guess it has trouble
inlining the VertexEmitter.vertex() calls because there are multiple
implementations, so reducing the number of calls and giving it a
chunkier function to JIT down helps? This is all conjecture because
I haven't figured out JitWatch yet :)

Anyways, this gives about a 9% improvement in my tests.