Don't calculate the exact vertex count on every VBO rebuild
This gives about a 3% improvement in VBO rebuild stress tests, for the
cost of a little more memory.

getVertexCount() was showing up heavy in my profiles. Changing it to a
simple upper bound calculation melts that time away. If there's a
max size 0.5 text scale monitor in the scene, the buffer will grow to
~3 MB. For comparison's sake, the images in the "blit" program were
already growing the buffer to ~2.1 MB.