Reduce cpu overhead in VBO backend.
- Avoid doubling up on state setup by adding a begin() and end() method
  to DirectVertexBuffer. In between calls to these methods, we can draw
  out of the vbo multiple times.

- Avoid setting up the VAO state every frame by using a vaoInitialised
  flag. My (possibly incorrect!) understanding of VAOs is that they're
  supposed to be used as persistent hold of all state centered around
  supplying vertices. We can set up the attribute vertex pointers once,
  stop binding/unbinding the vertex buffer, and keep the element array
  binding in the VAO state. I don't know why Mojang's version redoes all
  this constantly.

- Finally, reduce setupRenderState() work by setting up glPolygonOffset
  ourselves.

These changes make an obvious difference in profiles, but are really
hard to make apparent in game, at least on my machine. I'm perhaps not
cpu bound enough. So, to push my cpu I made a world with 10,240 1x1
monitors, and saw a savings of ~7ms per frame (17fps -> 21fps). That
is so far from a real world test though, that I wonder if any of this
is valuable to users?