Fix rendering glitches when index buffer resizes
When an index buffer grows beyond the 2^16 boundary, the type is
promoted from a SHORT to an INT. However, this type is not updated in
any of the existing DirectVertexBuffers, so we mis-interpret the index
buffer and render garbage.

To fix this, we now store the index buffer object instead (and also
acquire it during upload), and then read the type from that (which is
guaranteed to be always correct).

This is a little more aligned to how 1.19 does things (though 1.19 does
bind during upload rather than first draw).