make computers use GLASS redstone logic once more,
This was set so that way normal redstone, would not conflict with CC Redstone, this got changed back to stone on accident by Patbox because appearntly some culling issues were arrising? I dont know, but for the sake of redstone sanity, changing it back to glass

I do not know the "sound" method for 1.18.2, but this will have to do and some people may just have to suffer with their computers breaking with the glass sound.